Added teaser image field and updated field exports
Until now, the first of the products images was used for representing
the products in teasers. This caused problems since we needed to have
image of different original aspect ration for teasers compared to what
is used in the gallery on the product page. Added an additional image
field for this purpose, and updated the product display and form field
group settings accordingly.

Additionally, updated the taxonomy term fields so that they are
displayed via the taxonomy term module field formatter. It was
previously set to use a custom field formatter, but now that we're
overriding the taxonomy term page with our own view we can use the
default field formatter.